Extract fixture partials to separate files
